@@ -3,59 +3,64 @@
   "version": "0.1.0",
   "private": true,
   "dependencies": {
-    "@aws-amplify/cli-extensibility-helper": "^3.0.0",
-    "@aws-amplify/ui-react": "^3.2.0",
-    "@fortawesome/fontawesome-svg-core": "^1.2.26",
-    "@fortawesome/free-solid-svg-icons": "^5.12.0",
-    "@fortawesome/react-fontawesome": "^0.1.8",
-    "@material-ui/core": "^4.11.0",
-    "@material-ui/icons": "^4.9.1",
-    "@material-ui/lab": "^4.0.0-alpha.56",
-    "@types/jest": "^29.2.4",
-    "@types/node": "^18.11.12",
-    "@types/react": "^17.0.52",
-    "@types/react-dom": "^18.0.9",
+    "@aws-amplify/cli-extensibility-helper": "^3.0.1",
+    "@aws-amplify/ui-react": "^4.4.3",
+    "@blueprintjs/core": "^4.17.8",
+    "@blueprintjs/icons": "^4.14.5",
+    "@emotion/react": "^11.10.6",
+    "@emotion/styled": "^11.10.6",
+    "@fortawesome/fontawesome-svg-core": "^6.4.0",
+    "@fortawesome/free-solid-svg-icons": "^6.4.0",
+    "@fortawesome/react-fontawesome": "^0.2.0",
+    "@mui/icons-material": "^5.11.16",
+    "@mui/lab": "^5.0.0-alpha.125",
+    "@mui/material": "^5.11.16",
+    "@mui/styles": "^5.11.16",
+    "@types/jest": "^29.5.0",
+    "@types/node": "^18.15.11",
+    "@types/react": "^18.0.32",
+    "@types/react-dom": "^18.0.11",
     "assign-deep": "^1.0.1",
-    "aws-amplify": "^4.3.28",
+    "aws-amplify": "^5.0.23",
     "aws-amplify-react": "^5.1.9",
-    "bootstrap": "^4.4.1",
-    "cypress-file-upload": "^4.1.1",
+    "bootstrap": "^5.2.3",
+    "cypress-file-upload": "^5.0.8",
     "deep-object-diff": "^1.1.9",
-    "flv.js": "^1.5.0",
-    "hls.js": "^0.13.2",
-    "jquery": "^3.5.1",
+    "flv.js": "^1.6.2",
+    "hls.js": "^1.3.5",
+    "html-react-parser": "^3.0.15",
+    "jquery": "^3.6.4",
     "js-levenshtein": "^1.1.6",
-    "mediaelement": "^4.2.16",
-    "mirador": "^3.3.0",
-    "mocha": "^9.2.1",
-    "mochawesome": "^7.0.1",
-    "npm-check-updates": "^16.6.0",
-    "patch-package": "^6.2.2",
-    "pdfjs-dist": "^3.4.120",
-    "query-string": "^6.9.0",
-    "react": "^16.13.1",
+    "mediaelement": "^6.0.3",
+    "mirador": "~3.3.0",
+    "mocha": "^10.2.0",
+    "mochawesome": "^7.1.3",
+    "npm-check-updates": "^16.10.0",
+    "patch-package": "^6.5.1",
+    "pdfjs-dist": "^2.14.305",
+    "query-string": "^8.1.0",
+    "react": "^18.2.0",
     "react-3d-viewer": "^1.0.12",
     "react-csv": "^2.2.2",
-    "react-dom": "^16.13.1",
-    "react-focus-lock": "^2.4.1",
+    "react-dom": "^18.2.0",
+    "react-focus-lock": "^2.9.4",
     "react-fontawesome": "^1.7.1",
-    "react-ga4": "^2.0.0",
-    "react-helmet": "^5.2.1",
-    "react-html-parser": "^2.0.2",
+    "react-ga4": "^2.1.0",
+    "react-helmet": "^6.1.0",
     "react-quill": "^2.0.0",
-    "react-router-dom": "^5.1.2",
-    "react-scripts": "5.0.0",
-    "react-share": "^4.4.0",
-    "react-slick": "^0.26.1",
-    "reactstrap": "^8.2.0",
-    "sanitize-html": "^2.4.0",
-    "sass": "^1.49.8",
-    "semantic-ui-css": "^2.4.1",
-    "semantic-ui-react": "^2.0.1",
+    "react-router-dom": "^6.10.0",
+    "react-scripts": "^5.0.1",
+    "react-share": "^4.4.1",
+    "react-slick": "^0.29.0",
+    "reactstrap": "^9.1.8",
+    "sanitize-html": "^2.10.0",
+    "sass": "^1.60.0",
+    "semantic-ui-css": "^2.5.0",
+    "semantic-ui-react": "^2.1.4",
     "slick-carousel": "^1.8.1",
-    "typescript": "^4.9.4",
-    "uuid": "^8.3.2",
-    "x3dom": "^1.8.1"
+    "typescript": "^5.0.3",
+    "uuid": "^9.0.0",
+    "x3dom": "^1.8.2"
   },
   "lint-staged": {
     "src/**/*.{js,jsx,ts,tsx,json,css,scss,md}": [
@@ -96,10 +101,10 @@
   },
   "devDependencies": {
     "cypress": "^5.6.0",
-    "cypress-localstorage-commands": "^2.2.0",
+    "cypress-localstorage-commands": "^2.2.2",
     "husky": "^3.1.0",
-    "lint-staged": "^9.5.0",
-    "prettier": "^1.19.1",
+    "lint-staged": "^13.2.0",
+    "prettier": "^2.8.7",
     "semantic-ui-css": "git+https://github.com/Semantic-Org/Semantic-UI-CSS.git"
   }
 }
