@@ -1,18 +1,26 @@
 import React from "react";
-import ReactDOM from "react-dom";
+import * as ReactDOMClient from 'react-dom/client';
+import { BrowserRouter } from 'react-router-dom';
 import "./css/index.scss";
 import "./css/colors.scss";
 import App from "./App";
 import * as serviceWorker from "./serviceWorker";
 
-import Amplify from "aws-amplify";
+import { Amplify } from "aws-amplify";
 import config from "./aws-exports";
 
 import "bootstrap/dist/css/bootstrap.css";
 
 Amplify.configure(config);
 
-ReactDOM.render(<App />, document.getElementById("root"));
+const container = document.getElementById('root');
+const root = ReactDOMClient.createRoot(container);
+
+root.render(
+  <BrowserRouter>
+    <App />
+  </BrowserRouter>, 
+);
 
 // If you want your app to work offline and load faster, you can change
 // unregister() to register() below. Note this comes with some pitfalls.
