@@ -4,13 +4,13 @@ describe('language_config.spec: Selecting English loads English results', () => 
     cy.wait(10000)
     cy.get('div.facet-fields')
       .find('div.category > div > h3 > button#category')
-      .click()
+      .click({force: true})
       .invoke('text')
       .should('equal', 'Category');
     cy.get('input#archive', { timeout: 2000 }).click();
     cy.get('div.facet-fields')
       .find('div.language > div > h3 > button#language')
-      .click()
+      .click({force: true})
       .invoke('text')
       .should('equal', 'Language');
     cy.get('input#en', { timeout: 2000 }).click();
@@ -20,7 +20,7 @@ describe('language_config.spec: Selecting English loads English results', () => 
 
   it('Items should now be English', () => {
     cy.get('div.gallery-item > div.card > a').first()
-      .click();
+      .click({force: true});
       cy.wait(5000);
     cy.url().should("include", "/archive/");
     cy.get('div.details-section-metadata > table[aria-label="Item Metadata"] tbody', { timeout: 5000 })
